[#71] refactor:보관함 페이지 메뉴 바 컴포넌트 분리
